Added start menu and winner to play again menu

+++ Labs/Tag/main.py
@@ -9,7 +9,6 @@ clock = pygame.time.Clock()
 WIDTH = 1080  # screen width
 HEIGHT = 608  # screen height
 screen = pygame.display.set_mode((WIDTH,HEIGHT))  # create screen object
-screen.fill('black')  # fill the surface with blue
 
 block_size = (25, 25)
 
@@ -20,58 +19,60 @@ player2 = Player(200, 400, block_size, 'blue', wasd_keys, False)
 
 platforms = pygame.sprite.Group()
 build_border(screen, block_size, platforms)
-build_platform(600, 500, block_size, 300, platforms, screen)
-build_platform(400, 400, block_size,  200, platforms, screen)
-build_platform(200, 300, block_size,  200, platforms, screen)
-build_platform(700, 200, block_size,  8000, platforms, screen)
+build_level(block_size, platforms, screen)
 
 game_timer = Timer(screen)
-#menu = Menu(screen)
 
 running = True
 play = True
+started = False
 
 while running:
 
-    game_events = pygame.event.get()
+    screen.fill('black')
 
     # poll for events
+    game_events = pygame.event.get()
+
     # pygame.QUIT event means the user clicked X to close your window
     for event in game_events:
         if event.type == pygame.QUIT:
             running = False
 
-    if play:
-        player1.update(platforms, player2)
-        player2.update(platforms, player1)
-        game_timer.update()
-
-        # RENDER YOUR GAME HERE
-        screen.fill('black')
-        player1.draw(screen)
-        player2.draw(screen)
-        platforms.draw(screen)
-        game_timer.draw()
-
-        play = game_timer.play
-
-    else:
-        screen.fill('black')
-        player1.kill()
-        player2.kill()
-        [platform.kill() for platform in platforms]
-        play = play_again(screen, game_events)
-        if play:
+    if not started:
+        started = start_menu(screen, game_events)
+        if started:
             player1 = Player(100, 400, block_size, 'green', arrow_keys, True)
             player2 = Player(200, 400, block_size, 'blue', wasd_keys, False)
             build_border(screen, block_size, platforms)
-            build_platform(600, 500, block_size, 300, platforms, screen)
-            build_platform(400, 400, block_size, 200, platforms, screen)
-            build_platform(200, 300, block_size, 200, platforms, screen)
-            build_platform(700, 200, block_size, 8000, platforms, screen)
+            build_level(block_size, platforms, screen)
             game_timer = Timer(screen)
-        # menu.update()
-        # menu.draw()
+
+    else:
+        if play:
+            player1.update(platforms, player2)
+            player2.update(platforms, player1)
+            game_timer.update()
+
+            # RENDER YOUR GAME HERE
+            player1.draw(screen)
+            player2.draw(screen)
+            platforms.draw(screen)
+            game_timer.draw()
+
+            play = game_timer.play
+
+        else:
+            play = play_again(screen, game_events, (player1.it, player2.it))
+            player1.kill()
+            player2.kill()
+            [platform.kill() for platform in platforms]
+            if play:
+                player1 = Player(100, 400, block_size, 'green', arrow_keys, True)
+                player2 = Player(200, 400, block_size, 'blue', wasd_keys, False)
+                build_border(screen, block_size, platforms)
+                build_level(block_size, platforms, screen)
+                game_timer = Timer(screen)
 
     # flip() the display to put your work on screen
     pygame.display.flip()
